requirements.txt
transformers==4.36.0
sentencepiece==0.1.99
accelerate==0.25.0
bitsandbytes==0.41.1
protobuf==3.20.3
konlpy==0.6.0
wandb==0.16.1
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.2
pyyaml==6.0.1
tqdm==4.66.1
mecab-python3==1.0.8
tokenizers==0.15.0
openpyxl==3.1.2
